requirements.txt
fastapi 
uvicorn 
pymongo 
python-dotenv
pydantic
pyjwt
authlib
requests
